调整

+++ mulspider-core/src/main/java/com/example/core/db/rocksdb/RocksStore.java
@@ -256,6 +256,9 @@ public class RocksStore {
             iterator.next();
         }
 
+        if (map.isEmpty())
+            return map;
+
         try {
             db.deleteRange(cfHandle, firstKey, lastKey);
         } catch (RocksDBException e) {
